Created new CloseTime feature (#12)
Added CreateTime and CloseTime Variables to model package.
Set CreateTime to automatically update and create on bug creation.
Set CloseTime to be null con creation.

Adjusted UpdateStatus to incorperate changing of CloseTime to set time to Now ate that intervale.

Built 4 Unit Tests
One to test if the CreateTime populates correctly.
One to test  that CloseTime sets to null correctly.
One to make sure that Close only changes when bug is set to close.
One to make sure that close populates correctly.